probe: Added a very simple probe plugin
This is a plugin that regularly probes the network by sending a random payment
to a random destination. These do not succeed, since we randomly selected the
`payment_hash`, but can still give us a lot of insight into the network
health.

The plugin stores its results in a `sqlite3` database for later inspection.

Signed-off-by: Christian Decker <decker.christian@gmail.com>