rebalance: return last rebalanceall_msg even when nothing was stopped
When the last `rebalanceall` run terminated silently in the background,
it is still useful to get the ending message of that run.

This change will make the `rebalancestop` method to return the same
'nothing to stop' message plus the rebalanceall_msg when there
was a `rebalanceall` somewhere in the past.