sauron: make tests more robust
cln-grpc was dying and taking the rest of lightningd with it
(because it's important) because it was trying to listen on port
9736 in multiple tests simlutaneously. this PR allows the sauron
tests to take advantage of the new plyn-testing feature that
assigns the grpc plugin a random unused port by default. sauron's
tests were previously overriding these values in its override
fixtures, and this PR corrects that. there is still some brittle-
ness around the `TEST_NETWORK` global variable, but setting it
in the override `LightningNode` fixture at the top of each test,
immediately before the call to init the superclass, appears to
make the tests pass, so we'll leave it like this for now.